Updated the teacher dashboard
Updated the teacher dashboard to show items in an unordered list of flex boxes.
Added buttons to flex boxes that link to /t/manage. Need to add an ID to that url to get the specified session. Needs back-end implementation to be functional.

The list is determined by the lessons object in state. This will need to be set from the dashboard component with the response of the teachers session data request.